requirements.txt
playwright
pytest-playwright
pytest
pytest-html
pytest-xdist
allure-pytest